refactor: rename MimeType to ContentType
