+++ tomcat/src/main/java/org/apache/coyote/http11/ResourceLocator.java
@@ -5,7 +5,7 @@ import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Objects;
-import org.apache.coyote.http11.response.spec.MimeType;
+import org.apache.coyote.http11.general.ContentType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -26,7 +26,7 @@ public class ResourceLocator {
             URL url = getClass().getResource(prefix + path);
             Path filePath = Path.of(Objects.requireNonNull(url).getPath());
             String extension = path.substring(path.lastIndexOf(".") + 1);
-            return new Resource(MimeType.of(extension), Files.readString(filePath));
+            return new Resource(ContentType.of(extension), Files.readString(filePath));
         } catch (IOException | NullPointerException e) {
             throw new IllegalArgumentException("파일을 찾을 수 없습니다. path = " + prefix + path);
         }
